requirements.txt
telebot
vk_api
requests
random
python-dotenv